[Aptos Data Client] Add a new interface file.
