+++ state-sync/state-sync-v2/data-streaming-service/src/stream_engine.rs
@@ -25,7 +25,7 @@ use crate::{
 };
 use aptos_data_client::{
     global_summary::{AdvertisedData, GlobalDataSummary},
-    ResponsePayload,
+    interface::ResponsePayload,
 };
 use aptos_id_generator::{IdGenerator, U64IdGenerator};
 use aptos_logger::prelude::*;
